tui: Move batch submenu actions
Interactive mode handles primary hunk actions, flow selection, asset
prompts, and batch management from one large module.

Batch menu code has its own prompts and command calls, but it sits beside
the main action loop. That keeps batch-management command imports scattered
through the dispatcher instead of giving the submenu a named adapter.

This commit addresses that by moving the batch submenu, its batch
selectors, and its batch command calls into tui.batch_menu while
interactive.py delegates the b action to handle_batch_menu.

The next commit will update the TUI tests to patch the extracted menu
module and guard the new import boundary.
